+++ apps/vr-tests/src/stories/DividerConverged.stories.tsx
@@ -1,18 +1,9 @@
 import * as React from 'react';
 import Screener, { Steps } from 'screener-storybook/src/screener';
 import { storiesOf } from '@storybook/react';
-import { mergeStyles } from '@fluentui/react';
 import { Divider } from '@fluentui/react-divider';
 
-const verticalStyles = mergeStyles({
-  height: '400px',
-});
-
-const horizontalStyles = mergeStyles({
-  width: '400px',
-});
-
-storiesOf('Divider Converged', module)
+storiesOf('Divider', module)
   .addDecorator(story => (
     <Screener steps={new Steps().snapshot('default', { cropTo: '.testWrapper' }).end()}>
       {story()}
@@ -21,7 +12,7 @@ storiesOf('Divider Converged', module)
   .addStory(
     'Horizontal without content',
     () => (
-      <div className={horizontalStyles}>
+      <div style={{ width: '400px' }}>
         <Divider />
       </div>
     ),
@@ -30,7 +21,7 @@ storiesOf('Divider Converged', module)
   .addStory(
     'Horizontal with content',
     () => (
-      <div className={horizontalStyles}>
+      <div style={{ width: '400px' }}>
         <Divider>Today</Divider>
       </div>
     ),
@@ -39,7 +30,7 @@ storiesOf('Divider Converged', module)
   .addStory(
     'Horizontal Start Aligned',
     () => (
-      <div className={horizontalStyles}>
+      <div style={{ width: '400px' }}>
         <Divider alignContent="start">Today</Divider>
       </div>
     ),
@@ -48,26 +39,26 @@ storiesOf('Divider Converged', module)
   .addStory(
     'Horizontal End Aligned',
     () => (
-      <div className={horizontalStyles}>
+      <div style={{ width: '400px' }}>
         <Divider alignContent="end">Today</Divider>
       </div>
     ),
     { includeRtl: true },
   )
   .addStory('Vertical Center Aligned', () => (
-    <div className={verticalStyles}>
+    <div style={{ height: '400px' }}>
       <Divider vertical>Today</Divider>
     </div>
   ))
   .addStory('Vertical Start Aligned', () => (
-    <div className={verticalStyles}>
+    <div style={{ height: '400px' }}>
       <Divider vertical alignContent="start">
         Today
       </Divider>
     </div>
   ))
   .addStory('Vertical End Aligned', () => (
-    <div className={verticalStyles}>
+    <div style={{ height: '400px' }}>
       <Divider vertical alignContent="end">
         Today
       </Divider>
@@ -76,7 +67,7 @@ storiesOf('Divider Converged', module)
   .addStory(
     'Appearance subtle',
     () => (
-      <div className={verticalStyles}>
+      <div style={{ width: '400px' }}>
         <Divider appearance="subtle">Today</Divider>
       </div>
     ),
@@ -85,7 +76,7 @@ storiesOf('Divider Converged', module)
   .addStory(
     'Appearance strong',
     () => (
-      <div className={verticalStyles}>
+      <div style={{ width: '400px' }}>
         <Divider appearance="strong">Today</Divider>
       </div>
     ),
@@ -94,19 +85,19 @@ storiesOf('Divider Converged', module)
   .addStory(
     'Appearance brand',
     () => (
-      <div className={verticalStyles}>
+      <div style={{ width: '400px' }}>
         <Divider appearance="brand">Today</Divider>
       </div>
     ),
     { includeHighContrast: true, includeDarkMode: true },
   )
   .addStory('Inset', () => (
-    <div className={verticalStyles}>
+    <div style={{ width: '400px' }}>
       <Divider inset>Today</Divider>
     </div>
   ))
   .addStory('Vertical inset', () => (
-    <div className={verticalStyles}>
+    <div style={{ height: '400px' }}>
       <Divider inset vertical>
         Today
       </Divider>
