+++ packages/example-data/etc/example-data.api.md
@@ -4,52 +4,6 @@
 
 ```ts
 
-// @public
-export const AvatarExamples: {
-    readonly size: readonly [20, 24, 28, 32, 36, 40, 48, 56, 64, 72, 96, 120, 128];
-    readonly nameAndImage: {
-        name: string;
-        image: string;
-    }[];
-    readonly name: string[];
-    readonly image: string[];
-    readonly badge: readonly [{
-        readonly status: "available";
-    }, {
-        readonly status: "away";
-    }, {
-        readonly status: "busy";
-    }, {
-        readonly status: "doNotDisturb";
-    }, {
-        readonly status: "offline";
-    }, {
-        readonly status: "outOfOffice";
-    }, {
-        readonly status: "available";
-        readonly outOfOffice: true;
-    }, {
-        readonly status: "away";
-        readonly outOfOffice: true;
-    }, {
-        readonly status: "busy";
-        readonly outOfOffice: true;
-    }, {
-        readonly status: "doNotDisturb";
-        readonly outOfOffice: true;
-    }, {
-        readonly status: "offline";
-        readonly outOfOffice: true;
-    }, {
-        readonly status: "outOfOffice";
-        readonly outOfOffice: true;
-    }];
-    readonly activeDisplay: readonly ["ring", "ring-shadow", "ring-glow", "shadow", "glow"];
-    readonly color: readonly ["neutral", "brand", "colorful"];
-    readonly namedColors: readonly ["darkRed", "cranberry", "red", "pumpkin", "peach", "marigold", "gold", "brass", "brown", "forest", "seafoam", "darkGreen", "lightTeal", "teal", "steel", "blue", "royalBlue", "cornflower", "navy", "lavender", "purple", "grape", "lilac", "pink", "magenta", "plum", "beige", "mink", "platinum", "anchor"];
-    readonly hexagon: string;
-};
-
 // @internal (undocumented)
 export function createGroups(groupCount: number, groupDepth: number, startIndex: number, itemsPerGroup: number, level?: number, key?: string, isCollapsed?: boolean): IExampleGroup[];
