+++ packages/example-data/src/index.ts
@@ -5,7 +5,6 @@
  * can easily be added to the live example editor.)
  */
 
-export * from './AvatarExamples';
 export * from './facepile';
 export * from './listItems';
 export * from './lorem';
