+++ packages/storybook/src/index.ts
@@ -1,2 +1,1 @@
 export * from './decorators/index';
-export { withStrictMode } from '@fluentui/react-storybook';
