+++ packages/storybook/src/knobs/useStrictMode.ts
@@ -1,0 +1,3 @@
+import { boolean } from '@storybook/addon-knobs';
+
+export const useStrictMode = () => boolean('Use React.StrictMode', false);
